Modify updateNode to only return a Node instance.
This changes `updateNode` to raise exceptions if the the the node does
not exist, or if the the request was malformed, and allows the view to
decide what kind of HTTP response to return based on the the exception.

Now, if the a the node cannot be found, the view will respond with a
404. If the the node name is not in the request path, the view will
response with a 400.